hypr: toggleGreyscale.sh -> toggleGreyscale.fish, using highp float instead of mediump float in shaders

+++ .config/hypr/shaders/blank.glsl
@@ -2,7 +2,7 @@
 // By MahouShoujoMivutilde
 // blank shader
 
-precision mediump float;
+precision highp float; // https://github.com/hyprwm/Hyprland/issues/1140#issuecomment-1950960827
 varying vec2 v_texcoord;
 uniform sampler2D tex;
